[REFACTOR][IR] Use PrimType for compiler dtypes
Unify compiler-facing PrimExpr dtype state around PrimType as the source of expression type truth, while keeping runtime/storage dtype values as raw DLDataType at ABI and storage boundaries.

This squashes the PrimType refactor, review/CI follow-ups, raw dtype boundary repairs, and the benchmark-confirmed rewrite hot-path recovery into one PR commit.

+++ python/tvm/topi/sort.py
@@ -110,7 +110,7 @@ def argsort(data, valid_count=None, axis=-1, is_ascend=1, dtype="float32"):
         f = tvm.compile(s, [data, out], "llvm")
         dev = tvm.cpu()
         tvm_data = tvm.runtime.tensor(np_data, dev)
-        tvm_out = tvm.runtime.tensor(np.zeros(dshape, dtype=data.dtype), dev)
+        tvm_out = tvm.runtime.tensor(np.zeros(dshape, dtype=data.dtype.dtype), dev)
         f(tvm_data, tvm_out)
     """
     data_buf = tvm.tirx.decl_buffer(
