+++ tests/python/tirx-base/test_tir_intrin.py
@@ -48,8 +48,8 @@ def test_nearbyint():
 
     dev = tvm.cpu(0)
     n = 10
-    a = tvm.runtime.tensor(np.random.uniform(high=100, size=n).astype(A.dtype), dev)
-    a_rounded = tvm.runtime.tensor(np.random.uniform(size=n).astype(A_rounded.dtype), dev)
+    a = tvm.runtime.tensor(np.random.uniform(high=100, size=n).astype(A.dtype.dtype), dev)
+    a_rounded = tvm.runtime.tensor(np.random.uniform(size=n).astype(A_rounded.dtype.dtype), dev)
     func(a, a_rounded)
     # Note that numpys rint rounds to nearest integer with
     # ties to halfway is broken by rounding to even.
@@ -125,8 +125,8 @@ def test_unary_intrin():
 
         dev = tvm.cpu(0)
         n = 10
-        a = tvm.runtime.tensor(np.random.uniform(0.1, 0.5, size=n).astype(A.dtype), dev)
-        b = tvm.runtime.tensor(np.random.uniform(size=n).astype(A.dtype), dev)
+        a = tvm.runtime.tensor(np.random.uniform(0.1, 0.5, size=n).astype(A.dtype.dtype), dev)
+        b = tvm.runtime.tensor(np.random.uniform(size=n).astype(A.dtype.dtype), dev)
         func(a, b)
         tvm.testing.assert_allclose(b.numpy(), np_func(a.numpy()), atol=atol, rtol=rtol)
 
@@ -140,7 +140,7 @@ def test_unary_intrin():
                     np.random.uniform(1.1, 2.0, size=n // 2),
                     np.random.uniform(-2.0, -1.1, size=n // 2),
                 ]
-            ).astype(A.dtype)
+            ).astype(A.dtype.dtype)
             a2 = tvm.runtime.tensor(out_np, dev)
             b2 = tvm.runtime.tensor(np.empty_like(out_np), dev)
             func(a2, b2)
@@ -148,7 +148,7 @@ def test_unary_intrin():
             assert np.all(np.isnan(b2.numpy()))
         if name == "exp":
             n = 8
-            out_np = np.random.randint(-20, 20, size=n).astype(A.dtype)
+            out_np = np.random.randint(-20, 20, size=n).astype(A.dtype.dtype)
             a2 = tvm.runtime.tensor(out_np, dev)
             b2 = tvm.runtime.tensor(np.empty_like(out_np), dev)
             func(a2, b2)
@@ -239,9 +239,9 @@ def test_binary_intrin():
 
         dev = tvm.cpu(0)
         n = 10
-        a = tvm.runtime.tensor(np.random.uniform(0, 1, size=n).astype(A.dtype), dev)
-        b = tvm.runtime.tensor(np.random.uniform(0, 1, size=n).astype(B.dtype), dev)
-        c = tvm.runtime.tensor(np.random.uniform(size=n).astype(A.dtype), dev)
+        a = tvm.runtime.tensor(np.random.uniform(0, 1, size=n).astype(A.dtype.dtype), dev)
+        b = tvm.runtime.tensor(np.random.uniform(0, 1, size=n).astype(B.dtype.dtype), dev)
+        c = tvm.runtime.tensor(np.random.uniform(size=n).astype(A.dtype.dtype), dev)
         func(a, b, c)
         tvm.testing.assert_allclose(c.numpy(), np_func(a.numpy(), b.numpy()), atol=1e-5, rtol=1e-5)
 
@@ -266,9 +266,9 @@ def test_ldexp():
 
     dev = tvm.cpu(0)
     n = 10
-    a = tvm.runtime.tensor(np.random.uniform(0, 1, size=n).astype(A.dtype), dev)
-    b = tvm.runtime.tensor(np.random.randint(0, 5, size=n).astype(B.dtype), dev)
-    c = tvm.runtime.tensor(np.random.uniform(size=n).astype(A.dtype), dev)
+    a = tvm.runtime.tensor(np.random.uniform(0, 1, size=n).astype(A.dtype.dtype), dev)
+    b = tvm.runtime.tensor(np.random.randint(0, 5, size=n).astype(B.dtype.dtype), dev)
+    c = tvm.runtime.tensor(np.random.uniform(size=n).astype(A.dtype.dtype), dev)
     func(a, b, c)
     tvm.testing.assert_allclose(c.numpy(), np.ldexp(a.numpy(), b.numpy()), atol=1e-5, rtol=1e-5)
